Use a double for PrecisionModel scale, avoiding overflows
Fixes #652

git-svn-id: http://svn.osgeo.org/geos/trunk@3907 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ include/geos/geom/BinaryOp.h
@@ -456,14 +456,14 @@ BinaryOp(const Geometry* g0, const Geometry *g1, BinOp _Op)
               << std::endl;
 #endif
 
-		long unsigned int maxScale = 1e16;
+		double maxScale = 1e16;
 
     // Don't use a scale bigger than the input one
     if ( g0scale && g0scale < maxScale ) maxScale = g0scale;
     if ( g1scale && g1scale < maxScale ) maxScale = g1scale;
 
 
-		for (long unsigned int scale=maxScale; scale >= 1; scale /= 10)
+		for (double scale=maxScale; scale >= 1; scale /= 10)
 		{
 			PrecisionModel pm(scale);
 			GeometryFactory gf(&pm);
